ref: extract useBlackJackState to separate hook file
